+++ saleor/graphql/webhook/mutations/webhook_trigger.py
@@ -15,6 +15,12 @@ from ....permission.auth_filters import AuthorizationFilters
 from ....webhook.error_codes import WebhookTriggerErrorCode
 from ....webhook.event_types import WebhookEventAsyncType
 from ....webhook.models import Webhook
+from ....webhook.transport.asynchronous.transport import (
+    create_deliveries_for_subscriptions,
+    generate_deferred_payloads,
+    send_webhook_request_async,
+)
+from ....webhook.transport.utils import prepare_deferred_payload_data
 from ...core import ResolveInfo
 from ...core.doc_category import DOC_CATEGORY_WEBHOOKS
 from ...core.mutations import BaseMutation
@@ -148,13 +154,6 @@ class WebhookTrigger(BaseMutation):
 
     @classmethod
     def perform_mutation(cls, _root, info, **data):
-        from ....webhook.transport.asynchronous.transport import (
-            create_deliveries_for_subscriptions,
-            generate_deferred_payloads,
-            send_webhook_request_async,
-        )
-        from ....webhook.transport.utils import prepare_deferred_payload_data
-
         event_type, object, webhook = cls.validate_input(info, **data)
         delivery = None
         requestor = get_user_or_app_from_context(info.context)
