+++ saleor/webhook/transport/synchronous/transport.py
@@ -14,6 +14,7 @@ from ....core.db.connection import allow_writer
 from ....core.models import EventDelivery, EventPayload
 from ....core.tracing import webhooks_opentracing_trace
 from ....core.utils import get_domain
+from ....core.utils.events import call_event
 from ....core.utils.url import sanitize_url_for_logging
 from ....graphql.webhook.subscription_payload import (
     generate_payload_from_subscription,
@@ -21,14 +22,18 @@ from ....graphql.webhook.subscription_payload import (
 )
 from ....graphql.webhook.subscription_types import WEBHOOK_TYPES_MAP
 from ....graphql.webhook.utils import get_pregenerated_subscription_payload
+from ....payment import PaymentError
+from ....payment.interface import TransactionActionData
 from ....payment.models import TransactionEvent
 from ....payment.utils import (
+    create_failed_transaction_event,
     create_transaction_event_from_request_and_webhook_response,
     recalculate_refundable_for_checkout,
 )
 from ... import observability
 from ...const import WEBHOOK_CACHE_DEFAULT_TIMEOUT
 from ...event_types import WebhookEventSyncType
+from ...payloads import generate_transaction_action_request_payload
 from ...utils import get_webhooks_for_event
 from .. import signature_for_payload
 from ..utils import (
@@ -387,3 +392,71 @@ def trigger_all_webhooks_sync(
         if parsed_response := parse_response(response_data):
             return parsed_response
     return None
+
+
+def trigger_transaction_request(
+    transaction_data: "TransactionActionData", event_type: str, requestor
+) -> None:
+    if not transaction_data.transaction_app_owner:
+        create_failed_transaction_event(
+            transaction_data.event,
+            cause=(
+                "Cannot process the action as the given transaction is not "
+                "attached to any app."
+            ),
+        )
+        recalculate_refundable_for_checkout(
+            transaction_data.transaction, transaction_data.event
+        )
+        return
+    webhook = get_webhooks_for_event(
+        event_type, apps_ids=[transaction_data.transaction_app_owner.pk]
+    ).first()
+    if not webhook:
+        create_failed_transaction_event(
+            transaction_data.event,
+            cause="Cannot find a webhook that can process the action.",
+        )
+        recalculate_refundable_for_checkout(
+            transaction_data.transaction, transaction_data.event
+        )
+        return
+
+    if webhook.subscription_query:
+        delivery = None
+        try:
+            delivery = create_delivery_for_subscription_sync_event(
+                event_type=event_type,
+                subscribable_object=transaction_data,
+                webhook=webhook,
+            )
+        except PaymentError as e:
+            logger.warning("Failed to create delivery for subscription webhook: %s", e)
+        if not delivery:
+            create_failed_transaction_event(
+                transaction_data.event,
+                cause="Cannot generate a payload for the action.",
+            )
+            recalculate_refundable_for_checkout(
+                transaction_data.transaction, transaction_data.event
+            )
+            return
+    else:
+        payload = generate_transaction_action_request_payload(
+            transaction_data, requestor
+        )
+        with allow_writer():
+            # Use transaction to ensure EventPayload and EventDelivery are created together, preventing inconsistent DB state.
+            with transaction.atomic():
+                event_payload = EventPayload.objects.create_with_payload_file(payload)
+                delivery = EventDelivery.objects.create(
+                    status=EventDeliveryStatus.PENDING,
+                    event_type=event_type,
+                    payload=event_payload,
+                    webhook=webhook,
+                )
+    call_event(
+        handle_transaction_request_task.delay,
+        delivery.id,
+        transaction_data.event.id,
+    )
